+++ extensions/unicode-symbols/CHANGELOG.md
@@ -7,6 +7,7 @@
   - Also added an indicator to show that this character is actually in a different set
 - Fix issue where the extension breaks on `Ancient Symbols` in List Mode ([#14562](https://github.com/raycast/extensions/issues/14562))
 - Improve search, wasn't properly filtering characters in Fuse.js
+- Update to `unidata16`, which uses the latest `16.0.0` Unicode data
 - Updated dependencies
 
 ## [Filter irrelevant characters] - 2024-05-21